WIP Refactor Input, IconInput, PasswordInput, SearchInput.
