+++ src/components/IconInput/index.js
@@ -1,0 +1,5 @@
+import { withTheme } from 'emotion-theming';
+
+import IconInput from './IconInput';
+
+export default withTheme(IconInput);
