--- src/components/IconInputWrapper/index.js
@@ -1,5 +1,0 @@
-import { withTheme } from 'emotion-theming';
-
-import IconInputWrapper from './IconInputWrapper';
-
-export default withTheme(IconInputWrapper);
