+++ src/components/SearchInput/SearchInput.js
@@ -3,36 +3,17 @@ import PropTypes from 'prop-types';
 import { cx } from 'react-emotion';
 
 import Input from '../Input';
-import IconInputWrapper from '../IconInputWrapper';
+import IconInput from '../IconInput';
 import SearchIcon from './search.svg';
 
 /**
  * SearchInput component for forms.
  */
-const SearchInput = ({ disabled, ...props }) => (
-  <IconInputWrapper
-    role="search"
-    icon={({ className, disabledClassName }) => (
-      <SearchIcon
-        {...{ disabled }}
-        className={cx(className, { [disabledClassName]: disabled })}
-      />
-    )}
-    input={({ className }) => <Input {...{ ...props, disabled, className }} />}
-  />
+const SearchInput = ({ ...props }) => (
+  <IconInput role="search" {...props}>
+    {({ className }) => <SearchIcon {...{ className }} />}
+  </IconInput>
 );
-
-SearchInput.propTypes = {
-  /**
-   * Triggers disabled styles on the component. This is also forwarded as
-   * attribute to the <input> element.
-   */
-  disabled: PropTypes.bool
-};
-
-SearchInput.defaultProps = {
-  disabled: false
-};
 
 /**
  * @component
