+++ src/components/TextArea/TextArea.js
@@ -11,10 +11,9 @@ const baseStyles = css`
 /**
  * TextArea component for forms.
  */
-// const TextArea = styled(Input.withComponent('textarea'))`
-//   ${baseStyles};
-// `;
-const TextArea = () => null;
+const TextArea = styled(Input.withComponent('textarea'))`
+  ${baseStyles};
+`;
 
 /**
  * @component
