+++ src/index.js
@@ -35,7 +35,7 @@ export { default as NotificationList } from './components/NotificationList';
 // Misc
 export { default as Badge } from './components/Badge';
 export { default as Card, CardHeader, CardFooter } from './components/Card';
-export { default as IconInputWrapper } from './components/IconInputWrapper';
+export { default as IconInput } from './components/IconInput';
 export { default as Image } from './components/Image';
 export { default as LoadingBar } from './components/LoadingBar';
 export { default as Price } from './components/Price';
